Fix/topup on startup (#80)
* Fix topup on startup

The PurschaseRequestListener has to be added before we start
to listen to new purchase requests on the storage. Else we
will receive the request to topup, but the handler is not there
to actually perform the topup.

* Rename PurschaseRequestListener to PurchaseRequestHandler

We had two classes with almost the same name. 
This one is not listening for new requests though.
it is actually handling the purchase request so I renamed it
to handler.

* Refactor Eventlistner class into handler