Update to changes in the Stripe 'cancel-subscription' method
Before the 'cancel_at_period_end' flag could be used to signal if
a subscription should be done up immediately. This has now been
replaced with a new flag called 'invoice_now'. This change in the
Stripe API has also reversed the logic such that:

  cancel_at_period_end = false
  invoice_now = true

causes the same behaviour. Default for 'invoice_now' is false.