Adds support for cases where the payment for a plan is not done immediately
This update also includes changes due to changes in the Stripe subscription
"cancel" method. Previously the flag "atIntervalEnd" was used to indicate
that outstanding payments should be charged at the end of the plan period.
This is now replaced with the flag "invoiceNow" which says that outstanding
payment should be charged right away.

Ref.: https://stripe.com/docs/api/subscriptions/cancel?lang=curl