Refactors the use of 'region-code' in the payment API to 'tax-region', which is a more accurate term when dealing with taxes
For the time being 'region-code' can be used as the value to methods
in the payment API where a 'tax-region' is requested. This will work
as long as the following is fullfilled:

   1) A customer belongs to only one region
   2) A region equals to a country and two-letter are used
   3) A 'tax-region' equals to a country

At some point we will have:

   - Customer belonging to multiple regions
   - Regions covering more than one country or part of a country
   - Tax regions covering parts of a country, f.ex. US states
   - Tax regions extending beyond a country

When this happens, there will no longer be a one-to-one mapping
between region-code and tax-region and some other method for finding
which tax-region(s) that should be applied when as customer buys
something.