Use an uuid value as charge-id on purchase of a free product
When purchasing a free product, Stripe don't add a charge-id for
the purchase. As a temporarily solution an UUID value is used
as the charge-id in such cases. Attempt at refunding a purchase
with an UUID as the charge-id will fail. But such cases should
be easily idetifiable as the purchase is listed with zero amount.