Adds delay loop to some direct calls to Stripe used in acceptance test
In some cases acceptance test that checks for correct updates
in Stripe, as a result of requests to Prime, might fail because
the check is done "too early". This update tries to work around
this by performing the check in Stripe multiple times with delay
in between before it gives up.