Fix TCC Timer issue when session is moved between gateways
Since the session can be moved from one OCS gateway instance to
another, we might end up with trailing timers on the instence
from which the session was moved. The TCC Timer will setback the
state of the session to IDLE which might not be true as it now
lives on another node. This will disable the timer. Since the
gateway does not store any reservation locally it is not needed
at this point.